修正 offset 计算

+++ phonedatatool/unpack/offset.go
@@ -15,6 +15,6 @@ func (op *IndexPartOffsetPart) Parse(reader *bytes.Reader) error {
 	if _, err := reader.Read(buf); err != nil {
 		return fmt.Errorf("failed to read: %v", err)
 	}
-	op.IndexPartOffset = int64(binary.LittleEndian.Uint64(buf))
+	op.IndexPartOffset = int64(binary.LittleEndian.Uint32(buf))
 	return nil
 }
